// src/app/shared/models/companyProfile.ts
export interface CompanyProfile {
  company_ticker: string;
  company_name: string;
  exchange: string;
  exchange_type: string;
  currency: string;
  industry: string;
  sector: string;
  isin: string;
  country: string;
  normalizer: number;
  created_date: string;
  am_uid: string;
  market_cap: number;
  price: number;
  timestamp: string;
  market_date: string;
  open: number;
  close: number;
  high: number;
  low: number;
  vwap: number;
  volume: number;
  change_percent: number;
  date: string;
  adtv: number;
  adtv5: number;
  adtv10: number;
  adtv20: number;
  adtv60: number;
  isoutlier: string;
  aadtv: number;
  aadtv5: number;
  aadtv10: number;
  aadtv20: number;
  aadtv60: number;
  aadtv120: number;
}